- Fixed a bug which caused out-of-bounds index exception when code was compiled as 64-bit

+++ MiscMerge/MiscMergeEngine.m
@@ -466,7 +466,7 @@ static NSMutableDictionary *globalSymbols = nil;
 "*/
 - (void)setMainObject:(id)anObject
 {
-    unsigned oldIndex = NSNotFound;
+    NSUInteger oldIndex = NSNotFound;
 
     if (currentObject != nil)
     {
